+++ deepinv/physics/forward.py
@@ -260,7 +260,20 @@ class Physics(torch.nn.Module):  # parent class for forward models
                     and hasattr(self, key)
                     and isinstance(value, torch.Tensor)
                 ):
-                    self.register_buffer(key, value)
+
+                    if isinstance(getattr(self, key), torch.Tensor):
+                        if value.device.type != getattr(self, key).device.type:
+                            warnings.warn(
+                                f"The provided tensor for parameter '{key}' is on a different device ({value.device}) than the current parameter device ({getattr(self, key).device}). The current device will be used.",
+                                stacklevel=2,
+                            )
+
+                    # Move `value` to the buffer's device before updating
+                    # regardless of where the `value` tensor is located.
+                    # Also performs type casting if necessary.
+                    # If getattr(self, key) is None, torch.Tensor.to will
+                    # ignore the call and just return the original tensor.
+                    setattr(self, key, value.to(getattr(self, key)))
 
     # NOTE: Physics instances can hold instances of torch.Generator as
     # (possibly nested) attributes and they cannot be copied using deepcopy
@@ -373,6 +386,7 @@ class LinearPhysics(Physics):
         is used for computing it, and this parameter fixes the relative tolerance of the least squares algorithm.
     :param str solver: least squares solver to use. Choose between `'CG'`, `'lsqr'`, `'BiCGStab'` and `'minres'`. See :func:`deepinv.optim.linear.least_squares` for more details.
     :param bool implicit_backward_solver: If `True`, uses implicit differentiation for computing gradients through the :meth:`deepinv.physics.LinearPhysics.A_dagger` and :meth:`deepinv.physics.LinearPhysics.prox_l2`, using :func:`deepinv.optim.linear.least_squares_implicit_backward` instead of :func:`deepinv.optim.linear.least_squares`. This can significantly reduce memory consumption, especially when using many iterations. If `False`, uses the standard autograd mechanism, which can be memory-intensive. Default is `True`.
+    :param torch.device, str device: cpu or cuda, every registered buffer and module parameters are recursively pushed onto the device during initialization.
 
     |sep|
 
@@ -436,6 +450,7 @@ class LinearPhysics(Physics):
         tol=1e-4,
         solver="lsqr",
         implicit_backward_solver: bool = True,
+        device: torch.device | str = "cpu",
         **kwargs,
     ):
         super().__init__(
@@ -456,6 +471,37 @@ class LinearPhysics(Physics):
             warnings.warn(
                 "Using implicit_backward_solver with a low number of iterations may produce inaccurate gradients during the backward pass. If you are not doing backpropagation through `A_dagger` or `prox_l2`, ignore this message. If you are training unfolded models, consider increasing max_iter."
             )
+
+        device_holder = torch.tensor(0.0, device=device)
+        self.register_buffer("_device_holder", device_holder, persistent=False)
+        # pushes all parameters/buffers to the specified device, including `noise_model`
+        self.to(device)
+
+    @property
+    def device(self) -> torch.device | str:
+        r"""
+        Returns the device where the physics parameters/buffers are stored.
+
+        :return: device of the physics parameters.
+        """
+        warnings.warn(
+            "Following torch.nn.Module's design, the 'device' attribute is deprecated and will be removed in a future version. To move the module's buffers/parameters to a different device, use the `to()` method."
+        )
+
+        return self._device_holder.device
+
+    @device.setter
+    def device(self, value: torch.device | str):
+        r"""
+        Sets the device where the physics parameters/buffers are stored.
+
+        :param device: device to which the physics parameters will be moved.
+        """
+        warnings.warn(
+            "Following torch.nn.Module's design, the 'device' attribute is deprecated and will be removed in a future version, i.e. doing `physics.device = device` will no longer work and throw an `AttributeError`. Use `physics.to(device)` instead."
+        )
+
+        self.to(value)
 
     def A_adjoint(self, y, **kwargs):
         r"""
@@ -964,7 +1010,8 @@ class DecomposablePhysics(LinearPhysics):
         from the `V_adjoint` function and the `img_size` parameter.
         This automatic adjoint is computed using automatic differentiation, which is slower than a closed form adjoint, and can
         have a larger memory footprint. If you want to use the automatic adjoint, you should set the `img_size` parameter.
-    :param torch.nn.parameter.Parameter, float params: Singular values of the transform
+    :param torch.nn.parameter.Parameter, float mask: Singular values of the transform
+    :param torch.device, str device: cpu or cuda, every registered buffer and module parameters are recursively pushed onto the device during initialization.
 
     |sep|
 
@@ -1002,9 +1049,10 @@ class DecomposablePhysics(LinearPhysics):
         U_adjoint=None,
         V=None,
         mask=1.0,
+        device: torch.device | str = "cpu",
         **kwargs,
     ):
-        super().__init__(**kwargs)
+        super().__init__(device=device, **kwargs)
 
         assert not (
             U is None and not (U_adjoint is None)
@@ -1026,6 +1074,8 @@ class DecomposablePhysics(LinearPhysics):
         mask = torch.tensor(mask) if not isinstance(mask, torch.Tensor) else mask
         self.img_size = img_size
         self.register_buffer("mask", mask)
+
+        self.to(device)
 
     def A(self, x, mask=None, **kwargs) -> Tensor:
         r"""
@@ -1210,6 +1260,7 @@ class Denoising(DecomposablePhysics):
     The linear operator is just the identity mapping :math:`A(x)=x`
 
     :param torch.nn.Module noise: noise distribution, e.g., :class:`deepinv.physics.GaussianNoise`, or a user-defined torch.nn.Module. By default, it is set to Gaussian noise with a standard deviation of 0.1.
+    :param torch.device, str device: cpu or cuda, every registered buffer and module parameters are recursively pushed onto the device during initialization.
 
     |sep|
 
@@ -1229,10 +1280,23 @@ class Denoising(DecomposablePhysics):
 
     """
 
-    def __init__(self, noise_model: NoiseModel | None = None, **kwargs):
+    def __init__(
+        self,
+        noise_model: NoiseModel | None = None,
+        device: str | torch.device = "cpu",
+        **kwargs,
+    ):
         if noise_model is None:
             noise_model = GaussianNoise(sigma=0.1)
-        super().__init__(noise_model=noise_model, **kwargs)
+
+        if noise_model.rng is not None:
+            if noise_model.rng.device != device:
+                warnings.warn(
+                    f"argument `device`={device} is different from the random generator device of the noise model, `noise_model.rng.device`={noise_model.rng.device}. This will likely lead to errors during execution. The device argument will be ignored in favor of `noise_model.rng.device`={noise_model.rng.device}."
+                )
+                device = noise_model.rng.device
+
+        super().__init__(noise_model=noise_model, device=device, **kwargs)
 
 
 def adjoint_function(A, input_size, device="cpu", dtype=torch.float):